// budgetFlow-ui/src/app/services/models/category-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CategoryResponse {
  categoryType?: 'EXPENSE' | 'INCOME';
  createdByUserId?: number;
  hexColor?: string;
  id?: number;
  name?: string;
  parentId?: number;
  subCategories?: Array<number>;
}